+++ src/main/java/io/github/sakurawald/core/auxiliary/CollectionUtil.java
@@ -1,0 +1,9 @@
+package io.github.sakurawald.core.auxiliary;
+
+import java.util.List;
+
+public class CollectionUtil {
+    public static <T> boolean validIndex(int index, List<T> list) {
+        return index >= 0 && index < list.size();
+    }
+}