+++ src/main/java/io/github/sakurawald/module/initializer/back/config/model/BackConfigModel.java
@@ -4,4 +4,5 @@ public class BackConfigModel {
     public double ignore_distance = 32d;
     public boolean enable_back_on_death = true;
     public boolean enable_back_on_teleport = true;
+    public int max_back_location_entries_to_save = 3;
 }