+++ src/main/java/io/github/sakurawald/module/initializer/back/config/model/BackSavedPositionModel.java
@@ -1,9 +1,0 @@
-package io.github.sakurawald.module.initializer.back.config.model;
-
-import io.github.sakurawald.core.structure.SpatialPose;
-
-import java.util.HashMap;
-
-public class BackSavedPositionModel {
-    public HashMap<String, SpatialPose> player2lastPos = new HashMap<>();
-}